+++ lib/bloc/experience/exp_cubit.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:rro_web/bloc/base/base_cubit.dart';
 import 'package:rro_web/model/experience/exp_model.dart';
 import 'package:rro_web/utils/json_converter.dart';
@@ -6,8 +5,8 @@ import 'package:rro_web/utils/json_converter.dart';
 class ExpCubit extends BaseCubit {
   List<ExpModel> _expList = [];
 
-  init(BuildContext context) async {
-    _expList = await JsonConverter.getExpModel(context);
+  init() async {
+    _expList = await JsonConverter.getExpModel();
     refresh();
   }