+++ lib/bloc/profile/pro_cubit.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:rro_web/bloc/base/base_cubit.dart';
 import 'package:rro_web/model/profile/profile_model.dart';
 
@@ -7,8 +6,8 @@ import '../../utils/json_converter.dart';
 class ProCubit extends BaseCubit {
   late ProfileModel _model;
 
-  init(BuildContext context) async {
-    _model = await JsonConverter.getProModel(context);
+  init() async {
+    _model = await JsonConverter.getProModel();
     refresh();
   }