+++ lib/constant/dashboard_constant.dart
@@ -4,9 +4,17 @@ class DashMenuConstant {
   static const int exp = 2;
   static const int project = 3;
 }
-const String avatarImg = "assets/images/ic_avatar.png"; 
 const String homeContent =
+    "Hello, My name is Reno Rizky Ollantino. "
     "I am a Mobile Developer with more than 5 years of experience. "
     "I have developed several android mobile applications using Java. "
     "I also have experience creating mobile, desktop, and web applications using Flutter "
     "for more than 3 years.";
+
+const String githubUrl = "https://github.com/bleuxstrife";
+const String whatsappUrl = "whatsapp://send?phone=+6281285360029";
+const String emailUrl = "mailto:reno.rizky.ollantino@gmail.com";
+const String telUrl = "tel:+6281285360029";
+const String cvUrl = "https://1drv.ms/b/s!ApWNiqWX77amlFLdqb4KRGnlDXJJ?e=5cjx3H"; 
+
+typedef ContactOnClicked = void Function();