+++ lib/gen/assets.gen.dart
@@ -1,0 +1,622 @@
+/// GENERATED CODE - DO NOT MODIFY BY HAND
+/// *****************************************************
+///  FlutterGen
+/// *****************************************************
+
+// coverage:ignore-file
+// ignore_for_file: type=lint
+// ignore_for_file: directives_ordering,unnecessary_import,implicit_dynamic_list_literal
+
+import 'package:flutter/widgets.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:flutter/services.dart';
+
+class $AssetsImagesGen {
+  const $AssetsImagesGen();
+
+  /// File path: assets/images/ic_app_16.png
+  AssetGenImage get icApp16 =>
+      const AssetGenImage('assets/images/ic_app_16.png');
+
+  /// File path: assets/images/ic_app_192.png
+  AssetGenImage get icApp192 =>
+      const AssetGenImage('assets/images/ic_app_192.png');
+
+  /// File path: assets/images/ic_app_512.png
+  AssetGenImage get icApp512 =>
+      const AssetGenImage('assets/images/ic_app_512.png');
+
+  /// File path: assets/images/ic_avatar.png
+  AssetGenImage get icAvatar =>
+      const AssetGenImage('assets/images/ic_avatar.png');
+
+  /// File path: assets/images/ic_github.svg
+  SvgGenImage get icGithub => const SvgGenImage('assets/images/ic_github.svg');
+
+  /// File path: assets/images/ic_wa.svg
+  SvgGenImage get icWa => const SvgGenImage('assets/images/ic_wa.svg');
+
+  $AssetsImagesMerdekaGen get merdeka => const $AssetsImagesMerdekaGen();
+  $AssetsImagesNemobGen get nemob => const $AssetsImagesNemobGen();
+  $AssetsImagesTrinusaGen get trinusa => const $AssetsImagesTrinusaGen();
+
+  /// List of all assets
+  List<dynamic> get values =>
+      [icApp16, icApp192, icApp512, icAvatar, icGithub, icWa];
+}
+
+class $AssetsImagesMerdekaGen {
+  const $AssetsImagesMerdekaGen();
+
+  /// File path: assets/images/merdeka/merdeka_001.png
+  AssetGenImage get merdeka001 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_001.png');
+
+  /// File path: assets/images/merdeka/merdeka_002.png
+  AssetGenImage get merdeka002 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_002.png');
+
+  /// File path: assets/images/merdeka/merdeka_003.png
+  AssetGenImage get merdeka003 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_003.png');
+
+  /// File path: assets/images/merdeka/merdeka_004.png
+  AssetGenImage get merdeka004 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_004.png');
+
+  /// File path: assets/images/merdeka/merdeka_005.png
+  AssetGenImage get merdeka005 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_005.png');
+
+  /// File path: assets/images/merdeka/merdeka_006.png
+  AssetGenImage get merdeka006 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_006.png');
+
+  /// File path: assets/images/merdeka/merdeka_007.png
+  AssetGenImage get merdeka007 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_007.png');
+
+  /// File path: assets/images/merdeka/merdeka_008.png
+  AssetGenImage get merdeka008 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_008.png');
+
+  /// File path: assets/images/merdeka/merdeka_009.png
+  AssetGenImage get merdeka009 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_009.png');
+
+  /// File path: assets/images/merdeka/merdeka_010.png
+  AssetGenImage get merdeka010 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_010.png');
+
+  /// File path: assets/images/merdeka/merdeka_011.png
+  AssetGenImage get merdeka011 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_011.png');
+
+  /// File path: assets/images/merdeka/merdeka_012.png
+  AssetGenImage get merdeka012 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_012.png');
+
+  /// File path: assets/images/merdeka/merdeka_013.png
+  AssetGenImage get merdeka013 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_013.png');
+
+  /// File path: assets/images/merdeka/merdeka_014.png
+  AssetGenImage get merdeka014 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_014.png');
+
+  /// File path: assets/images/merdeka/merdeka_015.png
+  AssetGenImage get merdeka015 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_015.png');
+
+  /// File path: assets/images/merdeka/merdeka_016.png
+  AssetGenImage get merdeka016 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_016.png');
+
+  /// File path: assets/images/merdeka/merdeka_017.png
+  AssetGenImage get merdeka017 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_017.png');
+
+  /// File path: assets/images/merdeka/merdeka_018.png
+  AssetGenImage get merdeka018 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_018.png');
+
+  /// File path: assets/images/merdeka/merdeka_019.png
+  AssetGenImage get merdeka019 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_019.png');
+
+  /// File path: assets/images/merdeka/merdeka_020.png
+  AssetGenImage get merdeka020 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_020.png');
+
+  /// File path: assets/images/merdeka/merdeka_021.png
+  AssetGenImage get merdeka021 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_021.png');
+
+  /// File path: assets/images/merdeka/merdeka_022.png
+  AssetGenImage get merdeka022 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_022.png');
+
+  /// File path: assets/images/merdeka/merdeka_023.png
+  AssetGenImage get merdeka023 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_023.png');
+
+  /// File path: assets/images/merdeka/merdeka_024.png
+  AssetGenImage get merdeka024 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_024.png');
+
+  /// File path: assets/images/merdeka/merdeka_025.png
+  AssetGenImage get merdeka025 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_025.png');
+
+  /// File path: assets/images/merdeka/merdeka_026.png
+  AssetGenImage get merdeka026 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_026.png');
+
+  /// File path: assets/images/merdeka/merdeka_027.png
+  AssetGenImage get merdeka027 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_027.png');
+
+  /// File path: assets/images/merdeka/merdeka_028.png
+  AssetGenImage get merdeka028 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_028.png');
+
+  /// File path: assets/images/merdeka/merdeka_029.png
+  AssetGenImage get merdeka029 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_029.png');
+
+  /// File path: assets/images/merdeka/merdeka_030.png
+  AssetGenImage get merdeka030 =>
+      const AssetGenImage('assets/images/merdeka/merdeka_030.png');
+
+  /// List of all assets
+  List<AssetGenImage> get values => [
+        merdeka001,
+        merdeka002,
+        merdeka003,
+        merdeka004,
+        merdeka005,
+        merdeka006,
+        merdeka007,
+        merdeka008,
+        merdeka009,
+        merdeka010,
+        merdeka011,
+        merdeka012,
+        merdeka013,
+        merdeka014,
+        merdeka015,
+        merdeka016,
+        merdeka017,
+        merdeka018,
+        merdeka019,
+        merdeka020,
+        merdeka021,
+        merdeka022,
+        merdeka023,
+        merdeka024,
+        merdeka025,
+        merdeka026,
+        merdeka027,
+        merdeka028,
+        merdeka029,
+        merdeka030
+      ];
+}
+
+class $AssetsImagesNemobGen {
+  const $AssetsImagesNemobGen();
+
+  $AssetsImagesNemobHomeGen get home => const $AssetsImagesNemobHomeGen();
+  $AssetsImagesNemobLoginGen get login => const $AssetsImagesNemobLoginGen();
+  $AssetsImagesNemobOrderGen get order => const $AssetsImagesNemobOrderGen();
+  $AssetsImagesNemobReservationGen get reservation =>
+      const $AssetsImagesNemobReservationGen();
+}
+
+class $AssetsImagesTrinusaGen {
+  const $AssetsImagesTrinusaGen();
+
+  $AssetsImagesTrinusaHomeGen get home => const $AssetsImagesTrinusaHomeGen();
+  $AssetsImagesTrinusaLoginGen get login =>
+      const $AssetsImagesTrinusaLoginGen();
+  $AssetsImagesTrinusaPblGen get pbl => const $AssetsImagesTrinusaPblGen();
+  $AssetsImagesTrinusaTrinusaGen get trinusa =>
+      const $AssetsImagesTrinusaTrinusaGen();
+  $AssetsImagesTrinusaTtiGen get tti => const $AssetsImagesTrinusaTtiGen();
+}
+
+class $AssetsImagesNemobHomeGen {
+  const $AssetsImagesNemobHomeGen();
+
+  /// File path: assets/images/nemob/home/nemob_home_001.jpg
+  AssetGenImage get nemobHome001 =>
+      const AssetGenImage('assets/images/nemob/home/nemob_home_001.jpg');
+
+  /// File path: assets/images/nemob/home/nemob_home_002.jpg
+  AssetGenImage get nemobHome002 =>
+      const AssetGenImage('assets/images/nemob/home/nemob_home_002.jpg');
+
+  /// File path: assets/images/nemob/home/nemob_home_003.jpg
+  AssetGenImage get nemobHome003 =>
+      const AssetGenImage('assets/images/nemob/home/nemob_home_003.jpg');
+
+  /// File path: assets/images/nemob/home/nemob_home_004.jpg
+  AssetGenImage get nemobHome004 =>
+      const AssetGenImage('assets/images/nemob/home/nemob_home_004.jpg');
+
+  /// List of all assets
+  List<AssetGenImage> get values =>
+      [nemobHome001, nemobHome002, nemobHome003, nemobHome004];
+}
+
+class $AssetsImagesNemobLoginGen {
+  const $AssetsImagesNemobLoginGen();
+
+  /// File path: assets/images/nemob/login/nemob_login_001.jpg
+  AssetGenImage get nemobLogin001 =>
+      const AssetGenImage('assets/images/nemob/login/nemob_login_001.jpg');
+
+  /// File path: assets/images/nemob/login/nemob_login_002.jpg
+  AssetGenImage get nemobLogin002 =>
+      const AssetGenImage('assets/images/nemob/login/nemob_login_002.jpg');
+
+  /// File path: assets/images/nemob/login/nemob_login_003.jpg
+  AssetGenImage get nemobLogin003 =>
+      const AssetGenImage('assets/images/nemob/login/nemob_login_003.jpg');
+
+  /// List of all assets
+  List<AssetGenImage> get values =>
+      [nemobLogin001, nemobLogin002, nemobLogin003];
+}
+
+class $AssetsImagesNemobOrderGen {
+  const $AssetsImagesNemobOrderGen();
+
+  /// File path: assets/images/nemob/order/nemob_order_001.jpg
+  AssetGenImage get nemobOrder001 =>
+      const AssetGenImage('assets/images/nemob/order/nemob_order_001.jpg');
+
+  /// File path: assets/images/nemob/order/nemob_order_002.jpg
+  AssetGenImage get nemobOrder002 =>
+      const AssetGenImage('assets/images/nemob/order/nemob_order_002.jpg');
+
+  /// File path: assets/images/nemob/order/nemob_order_003.jpg
+  AssetGenImage get nemobOrder003 =>
+      const AssetGenImage('assets/images/nemob/order/nemob_order_003.jpg');
+
+  /// File path: assets/images/nemob/order/nemob_order_004.jpg
+  AssetGenImage get nemobOrder004 =>
+      const AssetGenImage('assets/images/nemob/order/nemob_order_004.jpg');
+
+  /// File path: assets/images/nemob/order/nemob_order_005.jpg
+  AssetGenImage get nemobOrder005 =>
+      const AssetGenImage('assets/images/nemob/order/nemob_order_005.jpg');
+
+  /// File path: assets/images/nemob/order/nemob_order_006.jpg
+  AssetGenImage get nemobOrder006 =>
+      const AssetGenImage('assets/images/nemob/order/nemob_order_006.jpg');
+
+  /// List of all assets
+  List<AssetGenImage> get values => [
+        nemobOrder001,
+        nemobOrder002,
+        nemobOrder003,
+        nemobOrder004,
+        nemobOrder005,
+        nemobOrder006
+      ];
+}
+
+class $AssetsImagesNemobReservationGen {
+  const $AssetsImagesNemobReservationGen();
+
+  /// File path: assets/images/nemob/reservation/nemob_reservation_001.jpg
+  AssetGenImage get nemobReservation001 => const AssetGenImage(
+      'assets/images/nemob/reservation/nemob_reservation_001.jpg');
+
+  /// File path: assets/images/nemob/reservation/nemob_reservation_002.jpg
+  AssetGenImage get nemobReservation002 => const AssetGenImage(
+      'assets/images/nemob/reservation/nemob_reservation_002.jpg');
+
+  /// File path: assets/images/nemob/reservation/nemob_reservation_003.jpg
+  AssetGenImage get nemobReservation003 => const AssetGenImage(
+      'assets/images/nemob/reservation/nemob_reservation_003.jpg');
+
+  /// File path: assets/images/nemob/reservation/nemob_reservation_004.jpg
+  AssetGenImage get nemobReservation004 => const AssetGenImage(
+      'assets/images/nemob/reservation/nemob_reservation_004.jpg');
+
+  /// File path: assets/images/nemob/reservation/nemob_reservation_005.jpg
+  AssetGenImage get nemobReservation005 => const AssetGenImage(
+      'assets/images/nemob/reservation/nemob_reservation_005.jpg');
+
+  /// File path: assets/images/nemob/reservation/nemob_reservation_006.jpg
+  AssetGenImage get nemobReservation006 => const AssetGenImage(
+      'assets/images/nemob/reservation/nemob_reservation_006.jpg');
+
+  /// File path: assets/images/nemob/reservation/nemob_reservation_007.jpg
+  AssetGenImage get nemobReservation007 => const AssetGenImage(
+      'assets/images/nemob/reservation/nemob_reservation_007.jpg');
+
+  /// List of all assets
+  List<AssetGenImage> get values => [
+        nemobReservation001,
+        nemobReservation002,
+        nemobReservation003,
+        nemobReservation004,
+        nemobReservation005,
+        nemobReservation006,
+        nemobReservation007
+      ];
+}
+
+class $AssetsImagesTrinusaHomeGen {
+  const $AssetsImagesTrinusaHomeGen();
+
+  /// File path: assets/images/trinusa/home/trinusa_home_001.PNG
+  AssetGenImage get trinusaHome001 =>
+      const AssetGenImage('assets/images/trinusa/home/trinusa_home_001.PNG');
+
+  /// File path: assets/images/trinusa/home/trinusa_home_002.PNG
+  AssetGenImage get trinusaHome002 =>
+      const AssetGenImage('assets/images/trinusa/home/trinusa_home_002.PNG');
+
+  /// File path: assets/images/trinusa/home/trinusa_home_003.PNG
+  AssetGenImage get trinusaHome003 =>
+      const AssetGenImage('assets/images/trinusa/home/trinusa_home_003.PNG');
+
+  /// File path: assets/images/trinusa/home/trinusa_home_004.PNG
+  AssetGenImage get trinusaHome004 =>
+      const AssetGenImage('assets/images/trinusa/home/trinusa_home_004.PNG');
+
+  /// File path: assets/images/trinusa/home/trinusa_home_005.PNG
+  AssetGenImage get trinusaHome005 =>
+      const AssetGenImage('assets/images/trinusa/home/trinusa_home_005.PNG');
+
+  /// List of all assets
+  List<AssetGenImage> get values => [
+        trinusaHome001,
+        trinusaHome002,
+        trinusaHome003,
+        trinusaHome004,
+        trinusaHome005
+      ];
+}
+
+class $AssetsImagesTrinusaLoginGen {
+  const $AssetsImagesTrinusaLoginGen();
+
+  /// File path: assets/images/trinusa/login/trinusa_login_001.PNG
+  AssetGenImage get trinusaLogin001 =>
+      const AssetGenImage('assets/images/trinusa/login/trinusa_login_001.PNG');
+
+  /// File path: assets/images/trinusa/login/trinusa_login_002.PNG
+  AssetGenImage get trinusaLogin002 =>
+      const AssetGenImage('assets/images/trinusa/login/trinusa_login_002.PNG');
+
+  /// List of all assets
+  List<AssetGenImage> get values => [trinusaLogin001, trinusaLogin002];
+}
+
+class $AssetsImagesTrinusaPblGen {
+  const $AssetsImagesTrinusaPblGen();
+
+  /// File path: assets/images/trinusa/pbl/trinusa_pbl_001.PNG
+  AssetGenImage get trinusaPbl001 =>
+      const AssetGenImage('assets/images/trinusa/pbl/trinusa_pbl_001.PNG');
+
+  /// File path: assets/images/trinusa/pbl/trinusa_pbl_002.PNG
+  AssetGenImage get trinusaPbl002 =>
+      const AssetGenImage('assets/images/trinusa/pbl/trinusa_pbl_002.PNG');
+
+  /// File path: assets/images/trinusa/pbl/trinusa_pbl_003.PNG
+  AssetGenImage get trinusaPbl003 =>
+      const AssetGenImage('assets/images/trinusa/pbl/trinusa_pbl_003.PNG');
+
+  /// File path: assets/images/trinusa/pbl/trinusa_pbl_004.PNG
+  AssetGenImage get trinusaPbl004 =>
+      const AssetGenImage('assets/images/trinusa/pbl/trinusa_pbl_004.PNG');
+
+  /// File path: assets/images/trinusa/pbl/trinusa_pbl_005.PNG
+  AssetGenImage get trinusaPbl005 =>
+      const AssetGenImage('assets/images/trinusa/pbl/trinusa_pbl_005.PNG');
+
+  /// File path: assets/images/trinusa/pbl/trinusa_pbl_006.PNG
+  AssetGenImage get trinusaPbl006 =>
+      const AssetGenImage('assets/images/trinusa/pbl/trinusa_pbl_006.PNG');
+
+  /// List of all assets
+  List<AssetGenImage> get values => [
+        trinusaPbl001,
+        trinusaPbl002,
+        trinusaPbl003,
+        trinusaPbl004,
+        trinusaPbl005,
+        trinusaPbl006
+      ];
+}
+
+class $AssetsImagesTrinusaTrinusaGen {
+  const $AssetsImagesTrinusaTrinusaGen();
+
+  /// File path: assets/images/trinusa/trinusa/trinusa_trinusa_001.PNG
+  AssetGenImage get trinusaTrinusa001 => const AssetGenImage(
+      'assets/images/trinusa/trinusa/trinusa_trinusa_001.PNG');
+
+  /// File path: assets/images/trinusa/trinusa/trinusa_trinusa_002.PNG
+  AssetGenImage get trinusaTrinusa002 => const AssetGenImage(
+      'assets/images/trinusa/trinusa/trinusa_trinusa_002.PNG');
+
+  /// File path: assets/images/trinusa/trinusa/trinusa_trinusa_003.PNG
+  AssetGenImage get trinusaTrinusa003 => const AssetGenImage(
+      'assets/images/trinusa/trinusa/trinusa_trinusa_003.PNG');
+
+  /// File path: assets/images/trinusa/trinusa/trinusa_trinusa_004.PNG
+  AssetGenImage get trinusaTrinusa004 => const AssetGenImage(
+      'assets/images/trinusa/trinusa/trinusa_trinusa_004.PNG');
+
+  /// File path: assets/images/trinusa/trinusa/trinusa_trinusa_005.PNG
+  AssetGenImage get trinusaTrinusa005 => const AssetGenImage(
+      'assets/images/trinusa/trinusa/trinusa_trinusa_005.PNG');
+
+  /// File path: assets/images/trinusa/trinusa/trinusa_trinusa_006.PNG
+  AssetGenImage get trinusaTrinusa006 => const AssetGenImage(
+      'assets/images/trinusa/trinusa/trinusa_trinusa_006.PNG');
+
+  /// File path: assets/images/trinusa/trinusa/trinusa_trinusa_007.PNG
+  AssetGenImage get trinusaTrinusa007 => const AssetGenImage(
+      'assets/images/trinusa/trinusa/trinusa_trinusa_007.PNG');
+
+  /// File path: assets/images/trinusa/trinusa/trinusa_trinusa_008.PNG
+  AssetGenImage get trinusaTrinusa008 => const AssetGenImage(
+      'assets/images/trinusa/trinusa/trinusa_trinusa_008.PNG');
+
+  /// List of all assets
+  List<AssetGenImage> get values => [
+        trinusaTrinusa001,
+        trinusaTrinusa002,
+        trinusaTrinusa003,
+        trinusaTrinusa004,
+        trinusaTrinusa005,
+        trinusaTrinusa006,
+        trinusaTrinusa007,
+        trinusaTrinusa008
+      ];
+}
+
+class $AssetsImagesTrinusaTtiGen {
+  const $AssetsImagesTrinusaTtiGen();
+
+  /// File path: assets/images/trinusa/tti/trinusa_tti_001.PNG
+  AssetGenImage get trinusaTti001 =>
+      const AssetGenImage('assets/images/trinusa/tti/trinusa_tti_001.PNG');
+
+  /// File path: assets/images/trinusa/tti/trinusa_tti_002.PNG
+  AssetGenImage get trinusaTti002 =>
+      const AssetGenImage('assets/images/trinusa/tti/trinusa_tti_002.PNG');
+
+  /// List of all assets
+  List<AssetGenImage> get values => [trinusaTti001, trinusaTti002];
+}
+
+class Assets {
+  Assets._();
+
+  static const $AssetsImagesGen images = $AssetsImagesGen();
+}
+
+class AssetGenImage {
+  const AssetGenImage(this._assetName);
+
+  final String _assetName;
+
+  Image image({
+    Key? key,
+    AssetBundle? bundle,
+    ImageFrameBuilder? frameBuilder,
+    ImageErrorWidgetBuilder? errorBuilder,
+    String? semanticLabel,
+    bool excludeFromSemantics = false,
+    double? scale,
+    double? width,
+    double? height,
+    Color? color,
+    Animation<double>? opacity,
+    BlendMode? colorBlendMode,
+    BoxFit? fit,
+    AlignmentGeometry alignment = Alignment.center,
+    ImageRepeat repeat = ImageRepeat.noRepeat,
+    Rect? centerSlice,
+    bool matchTextDirection = false,
+    bool gaplessPlayback = false,
+    bool isAntiAlias = false,
+    String? package,
+    FilterQuality filterQuality = FilterQuality.low,
+    int? cacheWidth,
+    int? cacheHeight,
+  }) {
+    return Image.asset(
+      _assetName,
+      key: key,
+      bundle: bundle,
+      frameBuilder: frameBuilder,
+      errorBuilder: errorBuilder,
+      semanticLabel: semanticLabel,
+      excludeFromSemantics: excludeFromSemantics,
+      scale: scale,
+      width: width,
+      height: height,
+      color: color,
+      opacity: opacity,
+      colorBlendMode: colorBlendMode,
+      fit: fit,
+      alignment: alignment,
+      repeat: repeat,
+      centerSlice: centerSlice,
+      matchTextDirection: matchTextDirection,
+      gaplessPlayback: gaplessPlayback,
+      isAntiAlias: isAntiAlias,
+      package: package,
+      filterQuality: filterQuality,
+      cacheWidth: cacheWidth,
+      cacheHeight: cacheHeight,
+    );
+  }
+
+  ImageProvider provider() => AssetImage(_assetName);
+
+  String get path => _assetName;
+
+  String get keyName => _assetName;
+}
+
+class SvgGenImage {
+  const SvgGenImage(this._assetName);
+
+  final String _assetName;
+
+  SvgPicture svg({
+    Key? key,
+    bool matchTextDirection = false,
+    AssetBundle? bundle,
+    String? package,
+    double? width,
+    double? height,
+    BoxFit fit = BoxFit.contain,
+    AlignmentGeometry alignment = Alignment.center,
+    bool allowDrawingOutsideViewBox = false,
+    WidgetBuilder? placeholderBuilder,
+    Color? color,
+    BlendMode colorBlendMode = BlendMode.srcIn,
+    String? semanticsLabel,
+    bool excludeFromSemantics = false,
+    Clip clipBehavior = Clip.hardEdge,
+    bool cacheColorFilter = false,
+    SvgTheme? theme,
+  }) {
+    return SvgPicture.asset(
+      _assetName,
+      key: key,
+      matchTextDirection: matchTextDirection,
+      bundle: bundle,
+      package: package,
+      width: width,
+      height: height,
+      fit: fit,
+      alignment: alignment,
+      allowDrawingOutsideViewBox: allowDrawingOutsideViewBox,
+      placeholderBuilder: placeholderBuilder,
+      color: color,
+      colorBlendMode: colorBlendMode,
+      semanticsLabel: semanticsLabel,
+      excludeFromSemantics: excludeFromSemantics,
+      clipBehavior: clipBehavior,
+      cacheColorFilter: cacheColorFilter,
+      theme: theme,
+    );
+  }
+
+  String get path => _assetName;
+
+  String get keyName => _assetName;
+}