+++ lib/ui/screen/dashboard/dash_desk_screen.dart
@@ -3,9 +3,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:rro_web/bloc/dashboard/dash_cubit.dart';
 import 'package:rro_web/bloc/experience/exp_cubit.dart';
+import 'package:rro_web/bloc/home/home_cubit.dart';
 import 'package:rro_web/bloc/profile/pro_cubit.dart';
 import 'package:rro_web/bloc/project/project_cubit.dart';
-import 'package:rro_web/constant/dashboard_constant.dart';
+import 'package:rro_web/gen/assets.gen.dart';
 import 'package:rro_web/utils/colors.dart';
 
 class DashDeskScreen extends StatelessWidget {
@@ -16,9 +17,10 @@ class DashDeskScreen extends StatelessWidget {
     return MultiBlocProvider(
       providers: [
         BlocProvider(create: (_) => DashCubit()),
-        BlocProvider(create: (_) => ExpCubit()..init(context)),
-        BlocProvider(create: (_) => ProCubit()..init(context)),
-        BlocProvider(create: (_) => ProjectCubit()..init(context))
+        BlocProvider(create: (_) => HomeCubit()),
+        BlocProvider(create: (_) => ExpCubit()..init()),
+        BlocProvider(create: (_) => ProCubit()..init()),
+        BlocProvider(create: (_) => ProjectCubit()..init())
       ],
       child: Builder(builder: (context) {
         final cubitWatch = context.watch<DashCubit>();
@@ -27,7 +29,7 @@ class DashDeskScreen extends StatelessWidget {
           padding: const EdgeInsets.all(16),
           child: CollapsibleSidebar(
               title: "RRO",
-              avatarImg: AssetImage(avatarImg),
+              avatarImg: AssetImage(Assets.images.icAvatar.path),
               backgroundColor: AppColors.primary[50]!,
               selectedIconBox: systemPrimaryColor,
               selectedIconColor: systemWhiteColor,